1. bugzilla bug 109324 DataSet element does not see Packages && bug 109327 BIRT error generated erroneously This is becasue when DataEngineImpl is initialized it creates a default scope that does not support importPackage keyword. Now we are creating the scope by newing an org.mozilla.javascript.ImporterTopLevel script object as scope so that importPackage keyword can be supported.
2. bugzilla bug 109698 In fetch method, the variable's value is defined in open method, but the value provided by afteropen method is expected.
In PreparedQuery, the code is executed in sequence as followed: dataSet.beforeOpen() -> executeOdiQuery() ->dataSet.afterOpen(). When using script data source, in method executeOdiQuery() the fetch method would be called so that the time point of fetch method is actually early than that of afterOpen() method. This is incorrect, however.

No we move the calling of afterOpen method from PreparedQuery to the executeOdiQuery() method of its children classes.So that ensure the afterOpen() method will be called immediately after the beforeOpen() method, or open() method(if exists).

3. Fix a bug in existing code. That bug causes that the global variable defined in two data source overlapped rather than seperate.

4. Add three new test cases to ScriptedDSTest