- Summary: CheckIn:Fix bugzilla bug 200370 OdaJndiName Property not checked - Bugzilla Bug (s) Resolved: [200370] OdaJndiName Property not checked
- Description:
This bug is occured when Odaurl is null and JndiNameUrl is not null.  Now we check these two values first, when they are both of null value, throw an exception. Otherwise, we will check JndiNameUrl first, if JndiNameUrl is null, then using Odaurl to try connecting.

Fixed in source.

- Tests Description:
Manual Test

- Files Edited:
/org.eclipse.birt.report.data.oda.jdbc/src/org/eclipse/birt/report/data/oda/jdbc/Connection.java
/org.eclipse.birt.report.data.oda.jdbc/src/org/eclipse/birt/report/data/oda/jdbc/JDBCDriverManager.java
/org.eclipse.birt.report.data.oda.jdbc/src/org/eclipse/birt/report/data/oda/i18n/ResourceConstants.java
/org.eclipse.birt.report.data.oda.jdbc/src/org/eclipse/birt/report/data/oda/i18n/Messages.properties

- Notes to Build Team:
N/A

- Notes to Developers:
N/A

- Notes to QA:
N/A

- Notes to Documentation:
N/A